Warnings.

+++ src/pk/ecc/ecc_verify_hash.c
@@ -46,7 +46,7 @@ int ecc_verify_hash_v2(const unsigned char *sig,
                                        int *stat,
                        const       ecc_key *key)
 {
-   if (opts->type < 0 || opts->type >= LTC_ARRAY_SIZE(s_ecc_verify_hash))
+   if (opts->type < 0 || (int)opts->type >= (int)LTC_ARRAY_SIZE(s_ecc_verify_hash))
       return CRYPT_PK_INVALID_TYPE;
    if (s_ecc_verify_hash[opts->type] == NULL)
       return CRYPT_PK_INVALID_TYPE;
